pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "obsidian_notes"
version = "0.1.0"
description = "A tool for processing and analyzing Obsidian notes"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "click",
    "inquirer",
    "llama-cpp-python",
    "openai",
    "python-dotenv",
    "pyyaml",
    "requests",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.4",
]
